don't tag ceph-ansible-3.0._ into luminous
CentOS generates a repository at
https://cbs.centos.org/repos/storage7-ceph-luminous-candidate/x86_64/os/Packages/
that is based on the "latest" build according to Koji.

Koji's definition for "latest" means "the last build to be tagged", not
"the highest NVR according to RPM's sorting rules".

In practice, this means that the following events would occur:

1) We would start with a ceph-ansible 3.1 build in the
   storage7-ceph-luminous-candidate yum repository,

2) ceph-ansible developers release a new 3.0 point release upstream,

3) We'd build and cross-tag that 3.0 point release into
   -luminous-candidate,

4) Koji would consider the 3.0 build to be the "latest" one, so it would
   essentially overwrite the older 3.1 build when regenerating the Yum
   repository.

The solution here is to avoid tagging 3.0 builds into
-luminous-candidate. We're going 3.1-only in Luminous now.